Core/Scripts : Add some Highmaul fixes - Bouncing Cleave targets melee players if no ennemy at ranged distance - Reduce the drop rate of epic stuff on Highmaul Conscript - Adjust the health of Vul'gor - Adjust the center pos of Brackenspore's beach - Positive mushrooms last 30s - Fix crash between npc_highmaul_volatile_anomalyAI::DamageTaken and spell_pri_power_word_shield_AuraScript::AfterAbsorb - Add missing template for Earthen Pillar - Players can't cast spells with Earthen Pillar between them - Crystalline Barrage cannot target melee players - Tectus's adds are spawned near to players - Pol energizes faster than Phemos - Enfeebling Roar correctly scales itself depending on target count - Blaze positions should be correctly synchronized with client - Ko'ragh is not stucked anymore if taunted during recharge - Suppression Field areatrigger now correctly affect targets inside - Volatile Anomaly cannot explode if the're silenced - Adjust Volatile Anomaly health and level - Add difficulty templates for Volatile Anomaly - Fix some server warnings - Remove deprecated code for AreaTriggers - Overflowing Energy should be paused during recharge - Improve the way Suppression Field is handled - Fix an issue with Caustic Energy - Add bonus loots on all boss (except Kargath, already added) - Berserker Rush cannot target tank or off tank - Frozen Core affects all allies within 8 yards - Fix multiple ticks of Crystalline Barrage

+++ src/server/game/Entities/Player/Player.cpp
@@ -21159,7 +21159,7 @@ bool Player::LoadFromDB(uint32 guid, SQLQueryHolder* holder, SQLQueryHolder* p_L
     if ((getMSTime() - l_StartTime) > 50)
     {
         sLog->outAshran("Player::LoadFromDB profiling =======");
-        for (int l_I = 0; l_I < l_Times.size(); l_I++)
+        for (int32 l_I = 0; l_I < (int32)l_Times.size(); l_I++)
             sLog->outAshran("Index [%u] : %u ms", l_I, l_Times[l_I]);
         sLog->outAshran("====================================");
     }
